daemon: move file `jsonrpc_` methods to an individual module
